Migrating `LookupTableFieldValueProviderFormContainer` to TS, replacing `FieldTypesStore` usage.
